Adding coordinates to annotation object now works

+++ include/util/json_container.hpp
@@ -124,7 +124,7 @@ using Value = mapbox::util::variant<String,
                                     Number,
                                     mapbox::util::recursive_wrapper<Object>,
                                     mapbox::util::recursive_wrapper<Array>,
-                                    std::tuple<float, float>,
+                                    std::tuple<double, double>,
                                     True,
                                     False,
                                     Null>;
